refactor(core): move SHA-256 lower hex helper to contracts
Why:
- SHA-256 lowercase hex formatting is used by CLI, infrastructure, Unity editor code, and application features, so infrastructure ownership forced cross-layer duplication.
- Review feedback also called out unnecessary byte-array allocation when finalizing incremental hashes.

What Changed:
- Redirected existing SHA-256 lower hex consumers to the Contracts cryptography helper and removed the Infrastructure-owned duplicate.
- Moved helper tests from Infrastructure tests to Contracts tests.
- Updated Utf8Sha256HashWriter to finalize IncrementalHash into a stackalloc span and to pass single characters as ReadOnlySpan<char>.

Impact:
- Digest formatting now has one shared implementation across layers.
- Incremental hash finalization avoids the previous GetHashAndReset byte-array allocation.
