+++ src/Ucli.Infrastructure/Index/FileContentHash.cs
@@ -1,4 +1,4 @@
-using MackySoft.Ucli.Infrastructure.Cryptography;
+using MackySoft.Ucli.Contracts.Cryptography;
 
 namespace MackySoft.Ucli.Infrastructure.Index;
