+++ src/Ucli.Infrastructure/Index/Scenes/SceneTreeLiteSourceHashCalculator.cs
@@ -1,5 +1,5 @@
 using System.Text;
-using MackySoft.Ucli.Infrastructure.Cryptography;
+using MackySoft.Ucli.Contracts.Cryptography;
 using MackySoft.Ucli.Infrastructure.Paths;
 
 namespace MackySoft.Ucli.Infrastructure.Index;
