+++ src/Ucli.Infrastructure/Project/UnityProjectFingerprintCalculator.cs
@@ -1,6 +1,6 @@
 using System.Runtime.InteropServices;
 using System.Text;
-using MackySoft.Ucli.Infrastructure.Cryptography;
+using MackySoft.Ucli.Contracts.Cryptography;
 using MackySoft.Ucli.Infrastructure.Paths;
 
 namespace MackySoft.Ucli.Infrastructure.Project;
