+++ src/Ucli/Features/Daemon/Supervisor/Bootstrap/SupervisorEndpointResolver.cs
@@ -1,6 +1,6 @@
 using System.Text;
+using MackySoft.Ucli.Contracts.Cryptography;
 using MackySoft.Ucli.Contracts.Ipc;
-using MackySoft.Ucli.Infrastructure.Cryptography;
 using MackySoft.Ucli.Infrastructure.Ipc;
 using MackySoft.Ucli.Infrastructure.Storage;
