+++ src/Ucli/UnityIntegration/Indexing/Core/FileReadIndexArtifactWriter.cs
@@ -1,8 +1,8 @@
 using System.Text;
 using MackySoft.Ucli.Application.Shared.Execution.ReadIndex;
+using MackySoft.Ucli.Contracts.Cryptography;
 using MackySoft.Ucli.Contracts.Index;
 using MackySoft.Ucli.Contracts.Json;
-using MackySoft.Ucli.Infrastructure.Cryptography;
 using MackySoft.Ucli.Infrastructure.Paths;
 using MackySoft.Ucli.Infrastructure.Storage;
