+++ tests/Ucli.Infrastructure.Tests/Storage/UcliStoragePathResolverContractTests.cs
@@ -1,7 +1,7 @@
 using System.Text;
 using MackySoft.Tests;
+using MackySoft.Ucli.Contracts.Cryptography;
 using MackySoft.Ucli.Contracts.Storage;
-using MackySoft.Ucli.Infrastructure.Cryptography;
 using MackySoft.Ucli.Infrastructure.Storage;
 
 namespace MackySoft.Ucli.Infrastructure.Tests.Storage;
