+++ tests/Ucli.Tests/UnityIntegration/Indexing/Core/FileReadIndexArtifactReaderTests.cs
@@ -1,7 +1,7 @@
 using System.Text;
 using MackySoft.Tests;
+using MackySoft.Ucli.Contracts.Cryptography;
 using MackySoft.Ucli.Contracts.Ipc;
-using MackySoft.Ucli.Infrastructure.Cryptography;
 using MackySoft.Ucli.Infrastructure.Storage;
 using MackySoft.Ucli.UnityIntegration.Indexing.Core;
